Rename output unformatted module
