+++ src/ECLlib/output/__init__.py
@@ -1,7 +1,13 @@
 """Output-related helpers for ECLlib."""
 
 from .egrid import EGRID_file
-from .unformatted import INIT_file, RFT_file, SMSPEC_file, UNRST_file, UNSMRY_file
+from .unformatted_files import (
+    INIT_file,
+    RFT_file,
+    SMSPEC_file,
+    UNRST_file,
+    UNSMRY_file,
+)
 from .textual import MSG_file, PRTX_file, PRT_file, text_file
 from .formatted import FUNRST_file, RSM_block, RSM_file, fmt_block, fmt_file
